% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_urls.R
\name{get_urls}
\alias{get_urls}
\title{Retrieve urls on google search}
\usage{
get_urls(search, how_many = 10)
}
\arguments{
\item{search}{A search string}

\item{how_many}{How many urls do you want to retrive}
}
\description{
This function provides a simple crawler to to retreive urls from
google search
}
\examples{
get_urls("machine learning", 10)

}
